adding login page.

+++ user-service/src/main/resources/db/migration/V3__create_user_payment_methods.sql
@@ -4,8 +4,8 @@ CREATE TABLE user_payment_methods (
     nickname      VARCHAR(50),
     card_type     VARCHAR(20)  NOT NULL,
     last_four     VARCHAR(4)   NOT NULL,
-    expiry_month  SMALLINT     NOT NULL,
-    expiry_year   SMALLINT     NOT NULL,
+    expiry_month  INTEGER      NOT NULL,
+    expiry_year   INTEGER      NOT NULL,
     is_default    BOOLEAN NOT NULL DEFAULT FALSE,
     created_at    TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
